# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/expense-report/{startDate}/{endDate}/{period}": {
            "get": {
                "summary": "Generate expense report",
                "operationId": "a39b2e41c1ad152069220dfa56aac32c",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "path",
                        "description": "Start date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "description": "End date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "path",
                        "description": "Report period (monthly, weekly, daily, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense report generated successfully"
                    }
                }
            }
        },
        "/income-report/{startDate}/{endDate}/{period}": {
            "get": {
                "summary": "Generate income report",
                "operationId": "7fe1fcc916cead4498641892426a0cf4",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "path",
                        "description": "Start date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "description": "End date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "path",
                        "description": "Report period (monthly, weekly, daily, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Income report generated successfully"
                    }
                }
            }
        },
        "/money-left-report/{startDate}/{endDate}/{period}": {
            "get": {
                "summary": "Generate money left report",
                "operationId": "52748d01e3a95526c993ad45780b8b7a",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "path",
                        "description": "Start date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "path",
                        "description": "End date of the report (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "path",
                        "description": "Report period (monthly, weekly, daily, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Money left report generated successfully"
                    }
                }
            }
        },
        "/search-expenses": {
            "post": {
                "summary": "Search expenses",
                "operationId": "1537099337998341cd38ca740e93dd1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "startDate": {
                                        "type": "string",
                                        "example": "2023-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "example": "2023-12-31"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "amountMax": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "expenseCategory": {
                                        "type": "string",
                                        "example": "food"
                                    },
                                    "expenseDescription": {
                                        "type": "string",
                                        "example": "restaurant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expenses search results"
                    }
                }
            }
        },
        "/search-incomes": {
            "post": {
                "summary": "Search incomes",
                "operationId": "704406e3f713e80e28bd09da6b1aa9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "startDate": {
                                        "type": "string",
                                        "example": "2023-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "example": "2023-12-31"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "amountMax": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "incomeType": {
                                        "type": "string",
                                        "example": "salary"
                                    },
                                    "incomeDescription": {
                                        "type": "string",
                                        "example": "bonus"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Incomes search results"
                    }
                }
            }
        },
...
    }
}